Refactor getting event type app metadata
